+++ src/common/node/robot/index.ts
@@ -1,0 +1,4 @@
+export * from './auto-hotkey';
+export * from './nircmd';
+export * from './robot';
+export * from './xdotool';
